LibPDF: Move font files into their own directory
