+++ Userland/Libraries/LibPDF/CMakeLists.txt
@@ -5,7 +5,8 @@ set(SOURCES
     Encoding.cpp
     Encryption.cpp
     Filter.cpp
-    Fonts.cpp
+    Fonts/PDFFont.cpp
+    Fonts/Type1Font.cpp
     ObjectDerivatives.cpp
     Parser.cpp
     Renderer.cpp
